Fixed handling of Exceptions while modifying a tracked view (issue #1)
Exception in tracked view may cause lock for prepared transactions not
to be cleaned up appropriately.
Exception during modification of a tracked view or another modification
listener currently immediately returns from the commit and no cleanup is
done. Instead we have to clean up the lock markers the prepared
transaction has on the modified objects. Otherwise later another TX
would get a stale object exception because of the current TX even if it
is finished a long time ago. Additionally even if one modification
causes an exception we should at least try to do the other
modifications. If multiple exceptions occur we should collect them as
suppressed exceptions.

Switched to JACIS v.1.43