Support optimistic locking without cloning
Provide methods (lockReadOnly...) to optimistically lock
an object to the current version.
This version is stored for the transaction.
In the prepare phase of the transaction it is checked
if the committed version has been changed,
that means the version of the committed entry
is higher than the locked version.
In this case a JacisStaleObjectException is thrown.